refactor: extract nav to client component
